refactor(di): reworked the load* methods to a different trait so as to fill code when Config is done
Refs #13
