+++ src/Di/Di.php
@@ -13,22 +13,18 @@ declare(strict_types=1);
 
 namespace Phalcon\Di;
 
-use Phalcon\Config\ConfigInterface;
 use Phalcon\Di\Exception\ServiceResolutionException;
 use Phalcon\Di\Traits\DiArrayAccessTrait;
 use Phalcon\Di\Traits\DiEventsTrait;
 use Phalcon\Di\Traits\DiExceptionsTrait;
 use Phalcon\Di\Traits\DiInstanceTrait;
+use Phalcon\Di\Traits\DiLoadTrait;
 use Phalcon\Events\ManagerInterface;
 use Phalcon\Events\Traits\EventsAwareTrait;
 
 use function is_object;
 use function lcfirst;
 use function substr;
-
-//use Phalcon\Config\Adapter\Php;
-//use Phalcon\Config\Adapter\Yaml;
-//use Phalcon\Config\ConfigInterface;
 
 /**
  * Phalcon\Di is a component that implements Dependency Injection/Service
@@ -79,6 +75,7 @@ class Di implements DiInterface
     use DiEventsTrait;
     use DiExceptionsTrait;
     use DiInstanceTrait;
+    use DiLoadTrait;
     use EventsAwareTrait;
 
     /**
@@ -340,103 +337,6 @@ class Di implements DiInterface
         }
 
         return $instance;
-    }
-
-    /**
-     * Loads services from a Config object.
-     */
-    protected function loadFromConfig(ConfigInterface $config): void
-    {
-        $services = $config->toArray();
-
-        foreach ($services as $name => $service) {
-            $this->set(
-                $name,
-                $service,
-                (bool) ($service['shared'] ?? false)
-            );
-        }
-    }
-
-    /**
-     * Loads services from a php config file.
-     *
-     * ```php
-     * $di->loadFromPhp("path/services.php");
-     * ```
-     *
-     * And the services can be specified in the file as:
-     *
-     * ```php
-     * return [
-     *      'myComponent' => [
-     *          'className' => '\Acme\Components\MyComponent',
-     *          'shared' => true,
-     *      ],
-     *      'group' => [
-     *          'className' => '\Acme\Group',
-     *          'arguments' => [
-     *              [
-     *                  'type' => 'service',
-     *                  'service' => 'myComponent',
-     *              ],
-     *          ],
-     *      ],
-     *      'user' => [
-     *          'className' => '\Acme\User',
-     *      ],
-     * ];
-     * ```
-     *
-     * @link https://docs.phalcon.io/en/latest/di
-     */
-    public function loadFromPhp(string $filePath): void
-    {
-        // $services = new Php(filePath);
-        $services = [];
-
-        $this->loadFromConfig($services);
-    }
-
-    /**
-     * Loads services from a yaml file.
-     *
-     * ```php
-     * $di->loadFromYaml(
-     *     "path/services.yaml",
-     *     [
-     *         "!approot" => function ($value) {
-     *             return dirname(__DIR__) . $value;
-     *         }
-     *     ]
-     * );
-     * ```
-     *
-     * And the services can be specified in the file as:
-     *
-     * ```php
-     * myComponent:
-     *     className: \Acme\Components\MyComponent
-     *     shared: true
-     *
-     * group:
-     *     className: \Acme\Group
-     *     arguments:
-     *         - type: service
-     *           name: myComponent
-     *
-     * user:
-     *    className: \Acme\User
-     * ```
-     *
-     * @link https://docs.phalcon.io/en/latest/di
-     */
-    public function loadFromYaml(string $filePath, array $callbacks = null): void
-    {
-        // $services = new Yaml($filePath, $callbacks);
-        $services = [];
-
-        $this->loadFromConfig($services);
     }
 
     /**
